Added Typescript, Vite
