+++ frontend/src/App.jsx
@@ -1,0 +1,13 @@
+import "App.css";
+import FullLayout from "layout/FullLayout";
+import { Box } from "@material-ui/core";
+
+const App = () => {
+  return (
+    <Box>
+      <FullLayout />
+    </Box>
+  );
+};
+
+export default App;
